Fix golangci-lint findings: errcheck, gocritic, prealloc, staticcheck, dupl
CI's lint job (golangci-lint run ./...) has been failing for a long time
undetected -- local verification only ever ran go build/go vet/go
test/gofmt, never golangci-lint itself, so nobody caught it despite the
lint job showing red on dozens of prior commits.

Real fixes, not suppressions:
- internal/agent/imageimport.go: check/discard cleanup errors
  (resp.Body.Close, os.Remove, tmp.Close) explicitly with `_ =` --
  best-effort cleanup, but no longer silently unchecked.
- internal/agent/imageimport_test.go, internal/csinode/controller_test.go:
  same treatment for test fixture writes/closes.
- internal/agent/agent.go: rewrite ensureRunning's if/else-if runtime
  state chain into an equivalent tagless switch (gocritic ifElseChain),
  no behavior change.
- internal/kaironctl/kaironctl_test.go: preallocate `names` with
  capacity len(got) (prealloc).
- internal/controller/machineset_test.go: the flagged
  `machineSetTemplateHash(t1) != machineSetTemplateHash(t1)` is NOT the
  copy-paste bug it looks like -- the test
  (TestMachineSetTemplateHashIsStableAndSensitiveToChange) deliberately
  hashes t1 twice to assert determinism, then separately compares t1 vs
  t2 to assert sensitivity to change. Left the assertion as-is with a
  targeted nolint explaining why.
- internal/migration/migration_test.go: rewrite an if/else-if phase
  comparison into a tagged switch (staticcheck QF1003).
- internal/nodeliveness/nodeliveness_test.go: drop the redundant
  embedded .Time selector -- model.MicroTime embeds time.Time, so
  RenewTime.After(...) already promotes (staticcheck QF1008).
- internal/csinode/statfs_linux.go: drop a redundant int64(...)
  conversion -- syscall.Statfs_t.Bsize is already int64 on
  linux/amd64 and linux/arm64 (unconvert).

Deduplicated the 5 flagged production-code dupl pairs with small,
narrow shared helpers (behavior-preserving refactors):
- internal/uiapi/qgadiag.go: handleQGAFirewallOpen/Close now both call
  a generic handleQGAFirewallToggle[Req], parameterized by request type
  and the open/close relay path segment.
- internal/uiapi/agentexec.go + agentfile.go: handleAgentExec and
  handleAgentGetFile now share relayGuestAgentRequest[Req, Resp] for the
  common decode-then-relay-then-respond tail.
- internal/kaironctl/kaironctl.go: cmdEditMachineSet/cmdEditMigrationPolicy
  now share editSpecFromFlags, which owns the common
  parse-only-passed-flags/build-spec/patch/print skeleton.
- internal/fluxvm/client.go + network.go: Client.Delete and
  DeleteNetworkGroup now share deleteIdempotent for the hand-rolled
  idempotent-404 DELETE (kept hand-rolled rather than routed through
  do(), same as before, to see the raw status code).
- internal/consoleproxy/server.go: handleQGAFirewallClose and
  handleAgentExec now share a generic relayFluxRequest[Req, Result, Resp]
  for the decode/call-with-timeout/encode shape.

+++ internal/csinode/statfs_linux.go
@@ -21,7 +21,7 @@ func statfs(path string) (VolumeStats, error) {
 	if err := syscall.Statfs(path, &buf); err != nil {
 		return VolumeStats{}, err
 	}
-	bsize := int64(buf.Bsize)
+	bsize := buf.Bsize
 	total := int64(buf.Blocks) * bsize
 	free := int64(buf.Bfree) * bsize
 	avail := int64(buf.Bavail) * bsize
